Connect BookmarksService with routes
Began the connection of routes using the BookmarksService object that was created.

There was some confusion as where to build this portion of the applicaiton as the course content did not make it clear whether or not we would be using the bookmarkRouter that we created a few lessons back.

Due to this an additiona app1.js was created to prevent the removal of other code that did not make sense for th4e project.